LauncherDecorator, fixed @symbol
* Separated LauncherDecorator into devoted class and made Run compatible
* Fixed @Symbol for MulticonfigWrapperOptions
