+++ src/main/java/com/synopsys/arc/jenkinsci/plugins/customtools/multiconfig/MulticonfigWrapperOptions.java
@@ -16,11 +16,13 @@
 package com.synopsys.arc.jenkinsci.plugins.customtools.multiconfig;
 
 import hudson.Extension;
+import hudson.ExtensionPoint;
 import hudson.matrix.MatrixProject;
 import java.io.Serializable;
 
 import hudson.model.Describable;
 import hudson.model.Descriptor;
+import jenkins.model.Jenkins;
 import org.jenkinsci.Symbol;
 import org.kohsuke.stapler.DataBoundConstructor;
 
@@ -28,7 +30,7 @@ import org.kohsuke.stapler.DataBoundConstructor;
  * Provides specific options for {@link MatrixProject}s.
  * @author Oleg Nenashev
  */
-public class MulticonfigWrapperOptions implements Serializable, Describable<MulticonfigWrapperOptions> {
+public class MulticonfigWrapperOptions implements Serializable, Describable<MulticonfigWrapperOptions>, ExtensionPoint {
     private final boolean skipMasterInstallation;
     public static final MulticonfigWrapperOptions DEFAULT = new MulticonfigWrapperOptions(false);
 
@@ -37,18 +39,23 @@ public class MulticonfigWrapperOptions implements Serializable, Describable<Mult
         this.skipMasterInstallation = skipInstallationOnMaster;
     }
 
+    @Deprecated
     public boolean isSkipMasterInstallation() {
         return skipMasterInstallation;
     }
 
+    public boolean isSkipInstallationOnMaster() {
+        return skipMasterInstallation;
+    }
+
     @Override
-    public Descriptor<MulticonfigWrapperOptions> getDescriptor() {
-        return new MulticonfigWrapperOptions.DescriptorImpl();
+    public MulticonfigWrapperOptionsDescriptor getDescriptor() {
+        return (MulticonfigWrapperOptionsDescriptor) Jenkins.getActiveInstance().getDescriptor(getClass());
     }
 
     @Symbol("customToolOptions")
     @Extension
-    public static class DescriptorImpl extends Descriptor<MulticonfigWrapperOptions> {
+    public static class MulticonfigWrapperOptionsDescriptor extends Descriptor<MulticonfigWrapperOptions> {
 
         @Override
         public String getDisplayName() {
